104 overriding mui styles without important (#112)
* Removed important's and added classes to handle hover and focus

* Unify look of pages

* Merge with dev

* Revert "Merge with dev"

This reverts commit db4eb09c36c16bc9aeced6d867e9911a3f2e7fd5.

* Fix merge conflicts in package-lock.json, install eslint again
