+++ frontend/package.json
@@ -52,6 +52,7 @@
         ]
     },
     "devDependencies": {
+        "@typescript-eslint/eslint-plugin": "^5.43.0",
         "cypress": "^11.0.1",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-cypress": "^2.12.1",
