+++ frontend/src/components/App.scss
@@ -4,3 +4,7 @@
     text-align: center;
     background-color: $background-color;
 }
+
+html {
+    background-color: $background-color;
+}
