+++ frontend/src/components/App.tsx
@@ -3,10 +3,14 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.scss';
 import NavBar from './NavBar';
 import PollCreation from './PollCreation';
+import Registration from './Registration';
+import Login from './Login';
 import BasicSnackbar from './BasicSnackbar';
 import PollAnswering from './PollAnswering';
 import AdminView from './AdminView';
 import PollResult from './PollResult';
+import { StyledEngineProvider } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
 
 function App() {
     const [open, setOpen] = useState(false);
@@ -40,41 +44,49 @@ function App() {
     };
 
     return (
-        <BrowserRouter>
-            <div className="App">
-                <NavBar></NavBar>
-
-                <Routes>
-                    <Route
-                        index
-                        element={
-                            <PollCreation
-                                showNotification={handleNotification}
-                                setPollId={handlePollId}
-                            ></PollCreation>
-                        }
-                    />
-                    <Route path="answer" element={<PollAnswering />} />
-                    <Route path="admin" element={<AdminView />} />
-                    <Route
-                        path="admin/result"
-                        element={
-                            <PollResult
-                                pollId={pollId}
-                                showNotification={handleNotification}
+        <>
+            <CssBaseline />
+            <StyledEngineProvider injectFirst>
+                <BrowserRouter>
+                    <div className="App">
+                        <NavBar></NavBar>
+                        <Routes>
+                            <Route
+                                index
+                                element={
+                                    <PollCreation
+                                        showNotification={handleNotification}
+                                        setPollId={handlePollId}
+                                    ></PollCreation>
+                                }
                             />
-                        }
-                    />
-                </Routes>
-
-                <BasicSnackbar
-                    open={open}
-                    onClose={handleClose}
-                    severity={notificationObj.severity}
-                    message={notificationObj.message}
-                />
-            </div>
-        </BrowserRouter>
+                            <Route path="/answer" element={<PollAnswering />} />
+                            <Route path="/admin" element={<AdminView />} />
+                            <Route path="/login" element={<Login />} />
+                            <Route
+                                path="/register"
+                                element={<Registration />}
+                            />
+                            <Route
+                                path="admin/result"
+                                element={
+                                    <PollResult
+                                        pollId={pollId}
+                                        showNotification={handleNotification}
+                                    />
+                                }
+                            />
+                        </Routes>
+                        <BasicSnackbar
+                            open={open}
+                            onClose={handleClose}
+                            severity={notificationObj.severity}
+                            message={notificationObj.message}
+                        />
+                    </div>
+                </BrowserRouter>
+            </StyledEngineProvider>
+        </>
     );
 }
