+++ frontend/src/components/BasicSnackbar.scss
@@ -1,11 +1,11 @@
 @import '../styles/colors';
 
 .MuiAlert-filledError {
-    background-color: $flamingo !important;
-    color: $black !important;
+    background-color: $flamingo;
+    color: $black;
 }
 
 .MuiAlert-filledSuccess {
-    background-color: $pear !important ;
-    color: $black !important;
+    background-color: $pear;
+    color: $black;
 }
