+++ frontend/src/components/Field.scss
@@ -2,47 +2,56 @@
 
 .user-field {
     .MuiInputBase-root {
-        width: 16.25rem !important;
+        width: 16.25rem;
         border-radius: 1.25rem;
         max-height: 2.5rem;
         margin-bottom: 1rem;
+
+        .MuiOutlinedInput-notchedOutline {
+            border: 1px solid $main-color;
+        }
     }
 }
 
-.MuiOutlinedInput-notchedOutline {
-    border: 1px solid $main-color !important;
+.user-field:last-child {
+    .MuiInputBase-root {
+        margin-bottom: 13px;
+    }
+}
+
+.label.Mui-focused {
+    color: $main-color;
 }
 
 .MuiInputLabel-shrink {
     position: absolute;
-    left: 2px;
 }
 
 .label {
-    color: $main-color !important;
-    font-size: medium !important;
+    color: $main-color;
+    font-size: medium;
 }
 
 .input-field-icon-login {
-    padding-left: 3.8rem;
+    margin-left: 3.8rem;
     margin-right: 0.5rem;
-    color: $black !important;
+    color: $black;
 }
 
 .input-field-icon-register {
-    padding-left: 2.7rem;
+    margin-left: 2.7rem;
     margin-right: 0.5rem;
-    color: $black !important;
+    color: $black;
 }
 
 .move-login {
-    padding-left: 5.8rem !important;
-    color: $placeholder-text-color !important;
-    font-size: 1rem !important;
+    margin-left: 5.8rem;
+    color: $placeholder-text-color;
+    font-size: 1rem;
 }
 
 .move-register {
-    padding-left: 4.7rem !important;
-    color: $placeholder-text-color !important;
-    font-size: 1rem !important;
+    margin-left: 4.7rem;
+    color: $placeholder-text-color;
+    font-size: 1rem;
 }
