+++ frontend/src/components/Login.tsx
@@ -1,7 +1,8 @@
-import { useState } from 'react';
-import { Typography, Container, Button, Link } from '@mui/material';
 import Field from './Field';
 import './LoginAndRegister.scss';
+import { useState } from 'react';
+import { Typography, Button, Link } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
 
 function Login(props: any) {
     const [username, setUsername] = useState('');
@@ -11,12 +12,18 @@ function Login(props: any) {
         index === 0 ? setUsername(value) : setPassword(value);
     };
 
+    const navigate = useNavigate();
+
+    const handleClick = () => {
+        navigate('/register');
+    };
+
     return (
-        <Container className="main-wrapper">
+        <div className="main-wrapper">
             <Typography variant="h4" className="page-heading">
                 Login
             </Typography>
-            <Container className="fields">
+            <div className="fields">
                 <Field
                     page="login"
                     text="Username"
@@ -39,15 +46,17 @@ function Login(props: any) {
                 >
                     Login
                 </Button>
-            </Container>
+            </div>
 
-            <Container className="texts">
+            <div className="texts">
                 <Typography className="bottom-text">
                     Don't have an account?
                 </Typography>
-                <Link className="bottom-text link">Register</Link>
-            </Container>
-        </Container>
+                <Link className="bottom-text link" onClick={handleClick}>
+                    Register
+                </Link>
+            </div>
+        </div>
     );
 }
