+++ frontend/src/components/LoginAndRegister.scss
@@ -7,62 +7,67 @@
 }
 
 .page-heading {
-    display: flex;
-    justify-content: center;
-    font-size: 2.5rem !important;
-    color: $main-color;
-    //font-weight: 600 !important;
+    font-weight: 600;
+    margin-top: 160px;
+    margin-bottom: 70px;
     .MuiTypography-root {
-        font-size: 2.5rem !important;
         @include for-phone-only {
             font-size: 2rem;
         }
     }
 }
 
-.MuiContainer-root {
-    margin-top: 180px;
-    display: flex !important;
+.main-wrapper,
+.fields {
+    margin-top: 13px;
+    display: flex;
     flex-direction: column;
     align-items: center;
 }
+.fields-wrapper {
+    border-radius: 20px;
+    border: 1px solid $black;
+    padding: 0rem 3.5rem;
+    background-color: $background-color;
+    margin-bottom: 25px;
+}
+.fields-wrapper:first-of-type {
+    margin-bottom: 1rem;
+}
 
 .login-btn {
-    border-radius: 20px !important;
-    width: 16.25rem !important;
-    color: $main-color !important;
-    background-color: $btn-background-color !important;
-    text-transform: none !important;
-    border: 2px solid $main-color !important;
-    margin: 0px !important;
-    margin-bottom: 1.25rem !important;
-    font-weight: 600 !important;
-    font-size: 1rem !important;
-    color: $black !important;
-    @include for-phone-only {
-        font-size: 2.5rem;
-    }
+    border-radius: 20px;
+    width: 16.25rem;
+    color: $main-color;
+    background-color: $sand;
+    text-transform: none;
+    border: 2px solid $main-color;
+    margin-top: 0px;
+    margin-bottom: 1.25rem;
+    font-weight: 600;
+    font-size: 1rem;
+    color: $black;
+}
+
+.login-btn:hover {
+    background-color: $sand;
+    border: 2px solid $main-color;
 }
 
 .bottom-text {
-    display: inline-block !important;
-    margin: 0rem 10px !important;
-    font-size: 1rem !important;
+    display: inline-block;
+    margin: 0rem 10px;
+    font-size: 1rem;
 }
 
 .texts {
-    margin-top: 0px !important;
-    display: inline-block !important;
-    margin-right: 20px !important;
-    font-size: 16px !important;
+    margin-top: 0px;
+    display: inline-block;
+    font-size: 16px;
 }
 
 .link {
     cursor: pointer;
     color: $link-color;
     text-decoration-color: $link-color;
-}
-
-.fields {
-    margin-top: 81px;
 }
