+++ frontend/src/components/ResultOption.scss
@@ -18,7 +18,7 @@
     }
 }
 
-.MuiSvgIcon-root {
+.MuiSvgIcon-root.check-icon {
     font-size: 1rem !important;
     position: relative;
     top: 3px;
