+++ frontend/src/components/Sidebar.scss
@@ -11,7 +11,7 @@
 }
 
 .list {
-    padding-top: 100px !important;
+    padding-top: 100px;
 }
 
 .page {
