+++ package.json
@@ -21,6 +21,7 @@
     "license": "MIT",
     "devDependencies": {
         "eslint": "^8.24.0",
+        "eslint-plugin-cypress": "^2.12.1",
         "husky": "^8.0.0",
         "lint-staged": "^13.0.3",
         "prettier": "^2.7.1"
